modified some logic. TODO: change reload logic with next instead prev field in each block (need to store also the first key of the rcu list to make the reload costs only O(n) directly from the device)

+++ src/helper.h
@@ -41,7 +41,7 @@
 struct block{
     short metadata;
     char data[MSG_MAX_SIZE];
-    int prev; //next block according to the order of the delivery of data
+    long prev; //next block according to the order of the delivery of data
 };
